Upgrade black to the latest.
The main code change is to change the command line arguments to
use type "click.UNPROCESSED" since there was a breaking change in
the new version of click that black uses.

Also this gets rid of ghpusher. We don't seem to use it and it's
very defunct, and depended on the old version of click.